requirements.txt
google_api_python_client == 2.165.0
Flask-Caching == 2.1.0
Flask-RESTful == 0.3.7
pycryptodome == 3.22.0
python-dotenv == 1.0.0
itsdangerous == 2.1.2
flask_restx == 1.2.0
MarkupSafe == 2.1.3
Flask-Cors == 4.0.0
aniso8601 == 6.0.0
Werkzeug == 2.2.2
urllib3 == 1.26.6
pyjwt == 2.10.1
Jinja2 == 3.1.2
Click == 7.1.2
Flask == 2.0.3
pytz == 2019.1
redis == 3.5.3
six == 1.12.0